refactor(meta): scripts update

+++ meta/gen_keystore.sh
@@ -1,26 +1,40 @@
-pushd /etc/pki/entitlement
-    PEM=$(ls -1 *-key.pem|sed 's/-key.pem//')
-popd
+#!/usr/bin/env bash
+
+if [[ -f ./keystore.p12 ]]
+then
+    exit 0
+fi
+
+PEMS=$(grep -e 'entitlement.*-key.pem' /etc/yum.repos.d/redhat.repo|sort|uniq|awk -F/ '{print $5}'|sed 's/-key\.pem//')
 
 PASS=$(cat /proc/sys/kernel/random/uuid)
-
-openssl pkcs12\
- -export\
- -in /etc/pki/entitlement/${PEM}.pem\
- -inkey /etc/pki/entitlement/${PEM}-key.pem\
- -name certificate_and_key\
- -out ./certificate_and_key.p12\
- -passout pass:${PASS}
-
-keytool\
- -importkeystore\
- -srckeystore ./certificate_and_key.p12\
- -srcstoretype PKCS12\
- -srcstorepass ${PASS}\
- -deststorepass ${PASS}\
- -destkeystore ./keystore.p12\
- -deststoretype PKCS12
-
-rm -f ./certificate_and_key.p12
-
 echo ${PASS} > ./keystore.pass
+
+for PEM in ${PEMS}
+do
+    if [[ -f /etc/pki/entitlement-host/${PEM}.pem ]]
+    then
+        ENTITLEMENT_PATH=/etc/pki/entitlement-host
+    else
+        ENTITLEMENT_PATH=/etc/pki/entitlement
+    fi
+
+    openssl pkcs12\
+     -export\
+     -in ${ENTITLEMENT_PATH}/${PEM}.pem\
+     -inkey ${ENTITLEMENT_PATH}/${PEM}-key.pem\
+     -name ${PEM}\
+     -out ./${PEM}.p12\
+     -passout pass:${PASS}
+
+    keytool\
+     -importkeystore\
+     -srckeystore ./${PEM}.p12\
+     -srcstoretype PKCS12\
+     -srcstorepass ${PASS}\
+     -deststorepass ${PASS}\
+     -destkeystore ./keystore.p12\
+     -deststoretype PKCS12
+
+    rm -vf ./${PEM}.p12
+done
